controller: son_master: handle scan trigger response
Preparative Commit.
Added a propagation of scan-trigger-request error response to the
dynamic-channel-selection-task. The following error responses will be
propagated to the DCS-task events:
CHANNEL_SCAN_OP_SCAN_IN_PROGRESS to SCAN_ALREADY_IN_PROGRESS
CHANNEL_SCAN_OP_ERROR to SCAN_TRIGGER_FAILED

This change is to allow the task to handle possible scan-trigger-request
errors and to retry in the case of failure due to
scan-already-in-progress error.

Signed-off-by: Adam Dov <adamx.dov@intel.com>