bwl: dwpal: send association frame to ap manager
Preparative commit.
Send the association frame to the AP manager, as part of the
STA_Connected event.

As mentioned in the previous commit message, the association frame we get
from the DWPAL contains the whole message, but the client capability
report should include only the frame body of the association request.

It is not clear from the specification what are the exact fields of the
message should be passed, but since certified units are passing only the
tagged parameters (after cleaning the Wi-Fi Alliance tag at from the end),
we do the same and pass the tagged params only. Otherwise, if we pass the
whole frame, the tests fail.

Therefore, we send only a substring of the original association frame we
get from the DWPAL, according to the substring range which was defined in
the previous commit.

Signed-off-by: Coral Malachi <coral.malachi@intel.com>