db: use supported channels for channel pool verification
In the set_channel_scan_pool function to validate the channel pool, we run a check to see if the
channel is supported. Currently the supported channel list is a static list.

Add is_channel_scan_pool_supported function to the DB.
Use the is_channel_scan_pool_supported function to validate channel pool.

Signed-off-by: itay elenzweig <itayx.elenzweig@intel.com>