tests: change test name caller format
As a follow up for the previous commit, remove the tcpdump flag if
statement from sniffer.start().

Second, add a dictionary of capture file locations for each test, later
to be accesed by tshark.

Third, when the sniffer is called, most of the times it will use test's
name string as-is, without the 'test_' prefix.
As a result - it makes it more readable to add the prefix once, instead of
trimming it a few times in sniffer.py.
Therefore, edit the test name's string in `test_flows.run_tests()` and handle
appropriately in `sniffer.start()`

Signed-off-by: Gal Wiener <gal.wiener@intel.com>