tests/sniffer: add tshark capture and parse method
Add a method parsing the pcap file into json using tshark,
deserializing it with python and applying the simpifier
method on each packet dictionary object.

This breaks the code as it uses a helper method which will be
implemented in the next commit, but is committed first as it makes
little sense to implement a method extending this one's functionality.

Signed-off-by: Gal Wiener <gal.wiener@intel.com>