son_slave_thread: trigger measurement request on sta connection
At the moment Associated STA Link Metrics collection is triggered by controller.
It is a vendor-specific message and it is sent just after STA connection.

Vendor-specific approach has disadvantage of not supporting non prplMesh contollers
and it is just redundant since there is a non-VS message for doing similar work.

We would like to replace controller-agent communication with non-VS messages,
but we would like to do it in tiny steps in order to not break anything.

This commit triggers statistics collection on association without
removing any code. Old flows need to be removed/refactored in future commits.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>