common: Add features to set_select_timeout function
Preparative commit.
We might want to temporarily change the select timeout and then reset it to initial value as
will be added to the follow-up commit.
Add "m_select_timeout_msec" member to socket thread to hold to initial select_timeout,
and "sticky" argument to "set_select_timeout" function to define if the given value is
temporarily or not.
The value of "0" will reset the value to the initial value.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>