beerocks_defines: replace SUPPORTED_CHANNELS_LENGTH with RADIO_CHANNELS_LENGTH
This is a preparative commit for "add preferred_channels to
ACTION_APMANAGER_JOINED_NOTIFICATION".

The SUPPORTED_CHANNELS_LENGTH is used for all radio channels, but the
name does not reflect that. Also the value of SUPPORTED_CHANNELS_LENGTH
is not enough to contain all possible channels and their bandwidths.

Change the SUPPORTED_CHANNELS_LENGTH with RADIO_CHANNELS_LENGTH
Increase the value of RADIO_CHANNELS_LENGTH to 128.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>