tlvf: move supported channels in ACTION_APMANAGER_JOINED_NOTIFICATION
After adding the preferred_channels to sNodeHostap and replacing the
misuses of the supported_channels, the supported_channels need to be
added to the CMDU so they will be sent to the son_slave.

Add supported_channels to the ACTION_APMANAGER_JOINED_NOTIFICATION.

This is a preparative commit for "tlvf: remove supported channels from
sNodeHostap".

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>