ci: cppcheck-new-issues.sh: remove the additional printed line
When a new issue is found, the first line of the next one was also
printed.

The -A option was used to display the caret that points the location
of the issue in the code.
However, it turns out the position of the caret is anyway incorrect
because the code is not on a newline.

Remove the -A option entirely, the exact position in the code can
still be retrieved from the original cppcheck job if needed.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>