bwl: add preferred_channels to RadioInfo
In the BWL the RadioInfo struct is used to store the radio's
infromation. The supported_channels are used for the ACS Report as well
as for the Basic Radio Capabilities.

To separate the two we need to use two containers, one for the ACS
Report (preferred_channels) and one of the Basic Radio Capabilities
(supported_channels).

This is a preparative commit for "bwl: update read_acs_report and
read_supported_channels".

Add preferred_channels to RadioInfo
Change radio channels containers to Array

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>