tools: cppcheck: output file names relative to rootdir
The issues in ci/cppcheck/cppcheck_existing_issues.txt contain
absolute paths (for some of them).

This is annoying because it prevents people to run
cppcheck-new-issues.sh locally.

Use the "rp" option so that file paths output by cppcheck are always
relative to the top-level folder.

The list of existing cppcheck issues is also updated to drop the
absolute paths.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>