ap_manager_thread: update radio channel list helper functions
In the ap_manager_thread there are two functions used to help with
copying and printing the preferred_channels (which used to be the
supported_channels). Since now we can use both supported_channels and
preferred_channels, the helper functions must be modified to support
both lists.

This is a preparative commit for "ap_manager_thread: use the new
supported channels in ACTION_APMANAGER_JOINED_NOTIFICATION".

Update channel list helper functions

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>