agent: backhaul_manager: collect link metrics for wireless interfaces
Refactor add_link_metrics() to be a class method instead of a static
function. Simplify the new method by moving out the code to create a
link metrics collector and collect the link metrics with it. New
method just adds TLVs to the Link Metric Response message.

Extend get_neighbor_links() to return wireless links as well as wired
links. Modify it to use the newer structs to report interfaces and
neighbors.

Update the Link Metric Query message handler to collect link metrics
for wireless links as well as for wired links.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>